requirements.txt
azure-core
azure-storage
chardet
dataset